% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiWisConnect.R
\name{exp_live_curl}
\alias{exp_live_curl}
\title{Test if an Experiment Hub is Live via Curl}
\usage{
exp_live_curl(hub, logPath = NULL)
}
\arguments{
\item{hub}{Character. Base URL of the Experiment Hub.}

\item{logPath}{Character or NULL. Path to a directory where logs will be written. Default is NULL (no logging).}
}
\value{
Logical. `TRUE` if the hub is reachable, `FALSE` otherwise.
}
\description{
This function checks if an Experiment Hub is reachable by querying a specific URL.
It first attempts to connect with SSL verification enabled. If that fails, it retries
without SSL verification. Optional logging records either the IP or any errors encountered.
}
\details{
The function constructs a URL to request the station list in JSON format.
It tries to retrieve the hub's IP address using `curl` first with SSL verification.
If the SSL-secured attempt fails, it retries without SSL verification and logs any errors.
If `logPath` is provided, the hub's IP or any errors are appended to `hub_ip.log`.
}
\examples{
\dontrun{
exp_live_curl("https://my.experimenthub.org", logPath = tempdir())
}

}
