% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiWisConnect.R
\name{get_kiwis_ts_list}
\alias{get_kiwis_ts_list}
\title{Retrieve a List of KIWIS Time Series}
\usage{
get_kiwis_ts_list(
  hub,
  station_no = NULL,
  dataType = NULL,
  return_fields = NULL,
  include_ts_name = NULL,
  exclude_ts_names = NULL,
  datasource = 1,
  logPath = NULL
)
}
\arguments{
\item{hub}{Character. URL of the KIWIS hub to query.}

\item{station_no}{Character or numeric vector. Station numbers to filter by. If NULL or empty, all stations are returned.}

\item{dataType}{Character vector. Parameter types to filter the time series. Defaults to all types ("*").}

\item{return_fields}{Character vector. Specific fields to return from the time series metadata. Defaults to commonly used fields.}

\item{include_ts_name}{Character vector. Time series names to include. Defaults to all ("*").}

\item{exclude_ts_names}{Character vector. Time series names to exclude from the results.}

\item{datasource}{Numeric or character. Identifier for the data source. Default is 1.}

\item{logPath}{Character. Optional path to log queries and responses.}
}
\value{
A data frame containing the requested time series metadata. Column names correspond to the fields specified in `return_fields`.
}
\description{
This function queries a KIWIS hub and retrieves a list of time series metadata for specified stations and data types. 
It allows filtering by station number, time series names, and parameter types, and supports returning specific fields.
}
\examples{
\dontrun{
hub_url <- "http://kiwis.kisters.de/KiWIS/KiWIS"
ts_list <- get_kiwis_ts_list(hub = hub_url, dataType = "Precip", datasource = 0)
head(ts_list)
}

}
