% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiWisConnect.R
\name{has_internet}
\alias{has_internet}
\title{Check Internet Connectivity}
\usage{
has_internet()
}
\value{
Logical. `TRUE` if the internet connection is available, `FALSE` otherwise.
}
\description{
This function checks if an internet connection is available by attempting 
to read a line from a known URL.
}
\details{
The function attempts to fetch the first line from "https://www.google.de".
Any warnings or errors during this process are suppressed. 
If the attempt fails, the function returns `FALSE`.
}
\examples{
if (has_internet()) {
  message("Internet connection available")
} else {
  message("No internet connection")
}

}
