% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiWisConnect.R
\name{write_zrxp_for_aquazis}
\alias{write_zrxp_for_aquazis}
\title{Write Dataset in ZRXP Format for AquaZIS}
\usage{
write_zrxp_for_aquazis(path, dataset)
}
\arguments{
\item{path}{Character. File path where the converted dataset will be saved.}

\item{dataset}{Character vector. The dataset to be converted and written to file. Typically obtained from KIWIS time series queries.}
}
\value{
None. The function writes the processed dataset to the specified file path.
}
\description{
This function converts a dataset to be compatible with the ZRXP format required by AquaZIS.
It renames specific columns (`ts_clientvalue6` → `EXTNUM1`, `ts_clientvalue7` → `EXTNUM2`), 
ensures UTF-8 encoding, removes any existing ZRXP headers, and writes the result to a file.
}
\examples{
\dontrun{
data_lines <- c("ts_clientvalue6,ts_clientvalue7,Value", "123,456,7.89")
write_zrxp_for_aquazis(path = "output.zrxp", dataset = data_lines)
}

}
